+++ test/react-web/client/graphql/queries-2.test.tsx
@@ -127,7 +127,7 @@ describe('queries', () => {
     class Container extends React.Component<any, any> {
       componentWillMount() {
         if (count === 1) {
-          expect(this.props.data.loading).toBe(true); // on remount
+          expect(this.props.data.loading).toBe.true; // on remount
           count++;
         }
       }
@@ -141,7 +141,7 @@ describe('queries', () => {
 
         if (count === 2) {
           // remounted data after fetch
-          expect(props.data.loading).toBe(false);
+          expect(props.data.loading).toBe.false;
           done();
         }
         count++;
@@ -176,11 +176,11 @@ describe('queries', () => {
     class Container extends React.Component<any, any> {
       render() {
         const { loading } = this.props.data;
-        if (count === 0) expect(loading).toBe(true);
-        if (count === 1) expect(loading).toBe(false);
-        if (count === 2) expect(loading).toBe(true);
+        if (count === 0) expect(loading).toBe.true;
+        if (count === 1) expect(loading).toBe.false;
+        if (count === 2) expect(loading).toBe.true;
         if (count === 3) {
-          expect(loading).toBe(false);
+          expect(loading).toBe.false;
           done();
         }
         count ++;
@@ -205,4 +205,72 @@ describe('queries', () => {
     // Prop update: fetch.
     setTimeout(() => render(variables2), 1000);
   });
+
+  it('correctly sets loading state on component with changed variables and unchanged result', (done) => {
+     const query = gql`
+       query remount($first: Int) { allPeople(first: $first) { people { name } } }
+     `;
+     const data = { allPeople: null };
+     const variables = { first: 1 };
+     const variables2 = { first: 2 };
+     const networkInterface = mockNetworkInterface(
+       { request: { query, variables }, result: { data }, delay: 10 },
+       { request: { query, variables: variables2 }, result: { data }, delay: 10 }
+     );
+     const client = new ApolloClient({ networkInterface });
+     let count = 0;
+
+     const connect = (component) : any => {
+       return class Container extends React.Component<any, any> {
+         constructor(props) {
+           super(props);
+
+           this.state = {
+             first: 1,
+           };
+           this.setFirst = this.setFirst.bind(this);
+         }
+
+         setFirst(first) {
+           this.setState({first});
+         }
+
+         render() {
+           return React.createElement(component, {
+             first: this.state.first,
+             setFirst: this.setFirst
+           });
+         }
+       }
+     }
+
+     @connect
+     @graphql(query, { options: ({ first }) => ({ variables: { first }})})
+     class Container extends React.Component<any, any> {
+       componentWillReceiveProps(props) {
+         if (count === 0) {
+           expect(props.data.loading).toBe.false; // has initial data
+           setTimeout(() => {
+             this.props.setFirst(2);
+           }, 5);
+         }
+
+         if (count === 1) {
+           expect(props.data.loading).toBe.true; // on variables change
+         }
+
+         if (count === 2) {
+           // new data after fetch
+           expect(props.data.loading).toBe.false;
+           done();
+         }
+         count++;
+       }
+       render() {
+         return null;
+       }
+     };
+
+     mount(<ProviderMock client={client}><Container /></ProviderMock>);
+   });
 });