+++ Changelog.md
@@ -8,6 +8,7 @@ Expect active development and potentially significant breaking changes in the `0
 - Feature: Move types to dev deps [#251](https://github.com/apollostack/react-apollo/pull/251)
 - Feature: New method for skipping queries which bypasses HOC internals [#253](https://github.com/apollostack/react-apollo/pull/253)
 - Feature: Integrated subscriptions! [#256](https://github.com/apollostack/react-apollo/pull/256)
+- Feature: Refactor loading state managment to use apollo-client fully. Reduces library size by ~50% [#211](https://github.com/apollostack/react-apollo/pull/211)
 
 ### v0.5.7